requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
Jinja2==3.1.6
click==8.1.1
itsdangerous==2.1.2
MarkupSafe==2.1.3
blinker==1.6.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
Pillow==10.0.0
requests==2.31.0
google-generativeai==0.3.2
PyJWT==2.8.0
itsdangerous==2.1.2
